Desktop-python/test1/requirements.txt
pyqt5>=5.15.0
pyserial>=3.5